# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal search and RAG tool — indexes data from multiple sources and provides unified search.",
        "title": "Nexus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/chats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the calling user's chats, ordered by updated_at desc.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "List chats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max chats to return (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of chats to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.listChatsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new empty chat owned by the requesting user. Both title and default_model are optional.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a chat",
                "parameters": [
                    {
                        "description": "Initial chat fields",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.createChatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_muty_nexus_internal_model.Chat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the chat plus its full message history. Owner-only; non-owners (including admins) get 404 to avoid leaking chat existence.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Get a chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.chatDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the chat and (via ON DELETE CASCADE) all its messages. Owner-only; non-owners get 404 to avoid leaking chat existence.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Delete a chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Applies a partial update (title and/or default_model). Owner-only; non-owners get 404 to avoid leaking chat existence.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Update a chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.updateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muty_nexus_internal_model.Chat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{id}/messages": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a user turn and streams the RAG orchestrator's events back as Server-Sent Events (retrieving, evidence, text, citation, tool_start/result, usage, title, done, error). Owner-only; non-owners get 404 to avoid leaking chat existence. The response is a long-lived text/event-stream, not a single JSON body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Send a chat message (SSE stream)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.postMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE event stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "503": {
                        "description": "RAG orchestrator not configured",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{id}/messages/{messageId}/feedback": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records a thumbs rating (\"up\"/\"down\", or null to clear) on a message. Owner-only; non-owners get 404 to avoid leaking chat existence.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Rate an assistant message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.messageFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_muty_nexus_internal_model.Chat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_model": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_muty_nexus_internal_model.ChatCitation": {
            "type": "object",
            "properties": {
                "cited_text": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "span_end": {
                    "type": "integer"
                },
                "span_start": {
                    "type": "integer"
                }
            }
        },
        "github_com_muty_nexus_internal_model.ChatListEntry": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_model": {
                    "type": "string"
                },
                "first_message_preview": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_muty_nexus_internal_model.ChatMessage": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "citations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatCitation"
                    }
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "description": "DurationMs is the wall-clock time the orchestrator's runTurn\ntook, in milliseconds. Persisted at orchestrator-side so the FE\ncan render a consistent label across the live (in-flight) view\nand the post-refresh persisted view. Nil for messages written\nbefore migration 019 and for user messages.",
                    "type": "integer"
                },
                "evidence": {
                    "description": "Evidence is the chunks the orchestrator retrieved for THIS turn\n(the same payload that streamed in EvEvidence). Persisting them\nalongside citations preserves the citation→source link end-to-end:\nChunkPreview.DocID is the OpenSearch chunk handle, so historical\nchats stay grounded — citations can render as numbered pills, the\nFE evidence rail repopulates, and any future feature can hop\n/api/documents/{id}, /related, /conversations, /blob from the\nstored handles.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChunkPreview"
                    }
                },
                "feedback": {
                    "description": "Feedback is the user's thumbs rating on an assistant message:\n\"up\", \"down\", or nil (no rating). Set via the message-feedback\nendpoint; used for Phase 7 telemetry. Always nil for user messages.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "rewritten_query": {
                    "description": "RewrittenQuery is the rewriter's normalised search query when the\nrewriter ran on this turn. Empty when the rewriter was disabled\nor this was the first user turn (rewriter only runs when prior\nassistant turns exist). Persisted so the FE phase strip can\nrepopulate \"Searching for: \u003crewritten\u003e\" on chat reload, and so\nthe Phase 7 eval harness knows which query actually hit OpenSearch.",
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatRole"
                },
                "seq": {
                    "type": "integer"
                },
                "skipped_retrieval": {
                    "description": "SkippedRetrieval is true when the rewriter judged the question\nanswerable from chat history alone (greetings, meta questions,\nhistory-only follow-ups). On these turns no OpenSearch call ran\nand Evidence is empty.",
                    "type": "boolean"
                },
                "stop_reason": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatToolCall"
                    }
                },
                "usage": {
                    "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatUsage"
                }
            }
        },
...
        "internal_api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.chatDetailResponse": {
            "type": "object",
            "properties": {
                "chat": {
                    "$ref": "#/definitions/github_com_muty_nexus_internal_model.Chat"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatMessage"
                    }
                }
            }
        },
...
        "internal_api.createChatRequest": {
            "type": "object",
            "properties": {
                "default_model": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.listChatsResponse": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muty_nexus_internal_model.ChatListEntry"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api.messageFeedbackRequest": {
            "type": "object",
            "properties": {
                "feedback": {
                    "type": "string"
                }
            }
        },
        "internal_api.postMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.updateChatRequest": {
            "type": "object",
            "properties": {
                "default_model": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by either a JWT from /auth/login or a personal access token (nexus_pat_...) from /tokens. The \"Bearer \" prefix is required; it is sent as the Authorization header verbatim.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}